+++ CHANGELOG.rst
@@ -1,9 +1,12 @@
 Change Log
 ==========
 
-[1.0.3] - 2019-03-28
+[1.0.3] - 2019-04-XX
 --------------------
 - [ADDED] Link to arXiv paper
+- [ADDED] Mexican-hat neighborhood distance weight
+- [ADDED] Possibility for different initialization modes
+- [CHANGED] Simplified initialization of estimators
 - [FIXED] URLs and styles in documentation
 - [FIXED] Colormap in Salinas example